Refactor package structure to move DroneController classes to portable_stock_ticker #83
