+++ src/main/java/de/theidler/create_mobile_packages/compat/jei/DroneControllerTransferHandler.java
@@ -5,8 +5,8 @@ import com.simibubi.create.content.logistics.packager.InventorySummary;
 import com.simibubi.create.content.logistics.stockTicker.CraftableBigItemStack;
 import com.simibubi.create.foundation.utility.CreateLang;
 import de.theidler.create_mobile_packages.index.CMPMenuTypes;
-import de.theidler.create_mobile_packages.items.drone_controller.DroneControllerMenu;
-import de.theidler.create_mobile_packages.items.drone_controller.DroneControllerScreen;
+import de.theidler.create_mobile_packages.items.portable_stock_ticker.DroneControllerMenu;
+import de.theidler.create_mobile_packages.items.portable_stock_ticker.DroneControllerScreen;
 import mezz.jei.api.gui.ingredient.IRecipeSlotsView;
 import mezz.jei.api.helpers.IJeiHelpers;
 import mezz.jei.api.recipe.RecipeIngredientRole;
