+++ src/main/java/de/theidler/create_mobile_packages/items/portable_stock_ticker/RequestStockUpdate.java
@@ -1,4 +1,4 @@
-package de.theidler.create_mobile_packages.items.drone_controller;
+package de.theidler.create_mobile_packages.items.portable_stock_ticker;
 
 import com.simibubi.create.foundation.networking.SimplePacketBase;
 import de.theidler.create_mobile_packages.index.CMPPackets;
@@ -7,7 +7,7 @@ import net.minecraft.server.level.ServerPlayer;
 import net.minecraftforge.network.NetworkEvent;
 import net.minecraftforge.network.PacketDistributor;
 
-import static de.theidler.create_mobile_packages.items.drone_controller.StockCheckingItem.getAccurateSummary;
+import static de.theidler.create_mobile_packages.items.portable_stock_ticker.StockCheckingItem.getAccurateSummary;
 
 public class RequestStockUpdate extends SimplePacketBase {
